README.md
This is a Vanilla JavaScript App which helps you calculate your monthly budget. You can add 'income' or 'expense' and it will be shown on the interface. The expenses will show what percentage of the total monthly budget they account for, separately and in total.

This app is done with the help of Jonas Schmedtmann's udemy course https://www.udemy.com/course/the-complete-javascript-course